dataproc_example/requirements.txt
# Google Cloud client libraries
google-cloud-dataproc>=5.0.0
google-cloud-storage>=2.0.0

# Apache Spark dependencies (for local development)
pyspark>=3.1.0

# Utility libraries
pandas>=1.3.0
numpy>=1.20.0

# Jupyter notebook (for local development)
jupyter>=1.0.0
notebook>=6.4.0

# Visualization libraries
matplotlib>=3.4.0
seaborn>=0.11.0

# Testing and development
pytest>=6.2.0
black>=21.5b2
flake8>=3.9.0
mypy>=0.812